+++ sdk/storage/azure_storage_blob/src/clients/mod.rs
@@ -3,6 +3,8 @@
 
 //! Clients used to communicate with Azure Blob Storage.
 
+use azure_core::http::{new_http_client, ClientOptions, HttpClientOptions, Transport};
+
 mod append_blob_client;
 mod blob_client;
 mod blob_container_client;
@@ -16,3 +18,13 @@ pub use blob_container_client::{BlobContainerClient, BlobContainerClientOptions}
 pub use blob_service_client::{BlobServiceClient, BlobServiceClientOptions};
 pub use block_blob_client::{BlockBlobClient, BlockBlobClientOptions};
 pub use page_blob_client::{PageBlobClient, PageBlobClientOptions};
+
+#[allow(clippy::needless_update)]
+fn apply_client_defaults(options: &mut ClientOptions) {
+    if options.transport.is_none() {
+        options.transport = Some(Transport::new(new_http_client(Some(HttpClientOptions {
+            automatic_decompression: false,
+            ..Default::default()
+        }))))
+    }
+}