+++ sdk/storage/azure_storage_blob/src/clients/page_blob_client.rs
@@ -60,6 +60,7 @@ impl PageBlobClient {
         options: Option<PageBlobClientOptions>,
     ) -> Result<Self> {
         let mut options = options.unwrap_or_default();
+        super::apply_client_defaults(&mut options.client_options);
         apply_storage_logging_defaults(&mut options.client_options);
 
         let storage_headers_policy = Arc::new(StorageHeadersPolicy);