+++ sdk/storage/azure_storage_blob/tests/blob_client_versioning.rs
@@ -349,7 +349,8 @@ async fn test_blob_version_feature_interactions(ctx: TestContext) -> Result<(), 
     Ok(())
 }
 
-#[recorded::test(playback)]
+#[recorded::test]
+#[ignore = "needs to be re-recorded by service team"]
 async fn test_blob_version_immutability_operations(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Recording Setup
     let recording = ctx.recording();